+++ app/src/main/java/com/scipath/becomeaking/BecomeAKing.java
@@ -5,16 +5,17 @@ import android.app.Application;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.scipath.becomeaking.contract.model.ICategory;
+import com.scipath.becomeaking.contract.model.IItem;
 import com.scipath.becomeaking.contract.model.IStats;
 import com.scipath.becomeaking.model.GameState;
 import com.scipath.becomeaking.manager.SaveManager;
 import com.scipath.becomeaking.model.Stats;
 import com.scipath.becomeaking.model.Personage;
 import com.scipath.becomeaking.model.enums.Stat;
+import com.scipath.becomeaking.model.item.Work;
 import com.scipath.becomeaking.view.fragment.DialogueFragment;
 
 import java.util.ArrayList;
-import java.util.function.Predicate;
 
 
 public class BecomeAKing extends Application {
@@ -103,6 +104,15 @@ public class BecomeAKing extends Application {
         gameState.personage.affectHealth(statBonuses.get(Stat.HealthPerDay));
         gameState.personage.affectReputation(statBonuses.get(Stat.ReputationPerDay));
         gameState.personage.affectMoney(statBonuses.get(Stat.CostPerDay));
+
+        ArrayList<ICategory> categories = gameState.categories;
+        for (int i = 0; i < categories.size(); i++) {
+            for (IItem work : categories.get(i).getItems()) {
+                if (work instanceof Work)
+                    work.setInteracted(false);
+            }
+        }
+        Work.refreshInteractionCounter();
     }
 
     public void checkPersonageForNegativeValues(AppCompatActivity activity) {