+++ app/src/main/java/com/scipath/becomeaking/adapter/CategoriesAdapter.java
@@ -19,6 +19,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.scipath.becomeaking.R;
 import com.scipath.becomeaking.contract.model.ICategory;
 import com.scipath.becomeaking.contract.model.IStats;
+import com.scipath.becomeaking.model.item.Work;
 import com.scipath.becomeaking.view.fragment.ItemsFragment;
 
 import java.util.ArrayList;
@@ -27,8 +28,8 @@ import java.util.ArrayList;
 public class CategoriesAdapter extends RecyclerView.Adapter<CategoriesAdapter.ViewHolder> {
 
     // Variables
-    private Context context;
-    private ArrayList<ICategory> categories;
+    private final Context context;
+    private final ArrayList<ICategory> categories;
 
 
     // Constructor
@@ -45,7 +46,6 @@ public class CategoriesAdapter extends RecyclerView.Adapter<CategoriesAdapter.Vi
 
         public ViewHolder(View view) {
             super(view);
-            // Define click listener for the ViewHolder's View
             layout = view.findViewById(R.id.category_layout);
         }
 
@@ -71,7 +71,6 @@ public class CategoriesAdapter extends RecyclerView.Adapter<CategoriesAdapter.Vi
     @NonNull
     @Override
     public CategoriesAdapter.ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int viewType) {
-        // Create a new view, which defines the UI of the list item
         View view = LayoutInflater.from(viewGroup.getContext())
                 .inflate(R.layout.layout_category, viewGroup, false);
 
@@ -84,16 +83,19 @@ public class CategoriesAdapter extends RecyclerView.Adapter<CategoriesAdapter.Vi
     public void onBindViewHolder(@NonNull CategoriesAdapter.ViewHolder viewHolder, @SuppressLint("RecyclerView") final int position) {
         // Get element from your dataset at this position and replace the
         // contents of the view with that element
-        ICategory ICategory = categories.get(position);
-        IStats stats = categories.get(position).getStats();
+        ICategory category = categories.get(position);
 
         // Setting values to views
-        viewHolder.getCategoryNameView().setText(ICategory.getName(context));
-        if (ICategory.getImageId()!= 0)
-            viewHolder.getCategoryImageView().setImageResource(ICategory.getImageId());
-        viewHolder.getCategoryImageView().setContentDescription(ICategory.getName(context));
-        viewHolder.getCategoryStatsView().setLayoutManager(new LinearLayoutManager(context));
-        viewHolder.getCategoryStatsView().setAdapter(new StatsAdapter(stats, context));
+        viewHolder.getCategoryNameView().setText(category.getNameId());
+        if (category.getImageId()!= 0)
+            viewHolder.getCategoryImageView().setImageResource(category.getImageId());
+
+        viewHolder.getCategoryImageView().setContentDescription(category.getName(context));
+        if (!(category.getItems().get(0) instanceof Work)) {
+            IStats stats = category.getStats();
+            viewHolder.getCategoryStatsView().setLayoutManager(new LinearLayoutManager(context));
+            viewHolder.getCategoryStatsView().setAdapter(new StatsAdapter(stats, context));
+        }
 
         viewHolder.getLayout().setOnClickListener(view -> {
             if (!categories.get(position).getItems().isEmpty()) {
@@ -111,7 +113,7 @@ public class CategoriesAdapter extends RecyclerView.Adapter<CategoriesAdapter.Vi
     }
 
 
-    // Return the size of your dataset (invoked by the layout manager)
+    // Return the size dataset (invoked by the layout manager)
     @Override
     public int getItemCount() {
         return categories.size();